feat: refactor to fully use outlet
